// src/types/auth.types.ts
export interface User {
  id: number;
  name: string;
  email: string;
  username?: string;
}

export interface LoginPayload {
  email: string;
  password: string;
}

export interface RegisterPayload {
  name: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}